refactor(core): split daily-logic + relocate cross-layer types (#1117)
Two coupled changes:

1. Move CapacityAssessment, ActionableIssue, ActionableIssueType,
   CompactActionableIssue, ActionMenuItem, ActionMenu from
   formatters/json.ts to core/types.ts. core/daily-logic.ts and
   test-utils.ts no longer reach into formatters/ — that was a
   backwards layer dependency. formatters/json.ts re-exports them so
   external importers don't break.

2. Split rendering out of daily-logic.ts:
   - core/daily-logic.ts: pure aggregations (applyStatusOverrides,
     toShelvedPRRef, groupPRsByRepo, computeRepoSignals,
     assessCapacity, collectActionableIssues, computeActionMenu).
   - commands/daily-render.ts: formatActionHint, formatBriefSummary,
     formatSummary, printDigest.
   - daily-logic.ts re-exports the rendering symbols so existing
     test imports keep working.

Verified:
- core/ no longer imports from formatters/ (grep clean).
- All 2077 core + 256 dashboard + 181 mcp tests pass without
  modification.
